kernels: move all locking bits to a `locking` module
